Set container image vars from OpenStackVersion
When setting the container image ansible variables in the generated
inventory, the OpenStackVersion resource is now checked. If an
OpenStackVersion resource exists, values are taken from that resource if
the vars are not already set in AnsibleVars by the user.

If an OpenStackVersion resource does not exist, the default values are
used instead, with the default values still being set from the
$RELATED_IMAGE* environment variables. Again, only if not explicitly set
by the user in AnsibleVars.

A watch on OpenStackVersion is also added to the NodeSet Manager, which
will trigger a reconcile for all NodeSets when OpenStackVersion changes.
The reconcile will cause the inventory to get regenerated.

The previous dnsMasqWatcher is renamed to genericWatcher, since both the
DNSMasq and OpenStackVersion watcher need to use the same logic of
generating reconcile requests for all NodeSets. genericWatcher is then
used as the watch function for both DNSMasq and OpenStackVersion
resources.

Signed-off-by: James Slagle <jslagle@redhat.com>