Add NodeSetHashes to Deployment.Status
The ConfigHash of the NodeSet is stored on Deployment.Status after the
deployment completes. This stored value is then used to set
NodeSet.Status.DeployedConfigHash.

This resolves an issue where NodeSet.Status.DeployedConfigHash was
always updated to ConfigHash, as long as one completed Deployment was
found, instead of only being updated when the actual hash has been
deployed.

Jira: OSPRH-5878
Signed-off-by: James Slagle <jslagle@redhat.com>