Add ContainerImageFields to OpenStackDataPlaneService
The ContainerImageFields on OpenStackDataPlaneService is a list of field
names that correspond to container fields on OpenStackVersion. When a
Service is deployed on a NodeSet, the names and URLs for the containers
are added to a ContainerImages map on Deployment and NodeSet.

The services under config/services are updated to set
ContainerImageFields with the container images they deploy.

Signed-off-by: James Slagle <jslagle@redhat.com>