Fix kuttl tests when running locally without webhooks
The tests that assert hashes rely on webhooks modifying the spec so that
the asserted hash value matches the calculated value. When running
locally without webhooks, these tests would fail. This commit adds the
relevant parts to the specs that the webhooks add so that the caluclated
hashes are equivalent when running locally.

Signed-off-by: James Slagle <jslagle@redhat.com>