Add EDPMServiceName to Service
Some services will have different names, but use the same ansible
content and be managing the same edpm service (such as in the case of
ovn and ovn-update).

The EDPMServiceName field, when set, will be used as the value for the
edpm_service_name ansible variable. This variable is used by
edpm-ansible content when generating the path of certs to mount into
containers, such that both services can use the same path
(/var/lib/openstack/certs/ovn).

Jira: OSPRH-6532
Signed-off-by: James Slagle <jslagle@redhat.com>