Remove config hashes from kuttl tests
We changed order of fields in the structs[1] that would change these
hashes. As we did not test the PR in prow kuttl job tests did
not fail. Now we need to remove these so that we can test with
PRs and merge [2].

Dropping `dataplane-deploy-nodehashes-test` as we've to remove the
hashes and there is no place we're asserting that hashes are different
after the edit. The test can be reworked and added later.

[1] https://github.com/openstack-k8s-operators/dataplane-operator/pull/758
[2] https://github.com/openshift/release/pull/51299