Add EDPMRoleServiceName to OpenstackDataPlaneServiceCert
EDPMRoleServiceName is used as the base dir for the mount path for the
cert. If not specified, OpenStackDataPlaneService.Spec.EDPMServiceName
is used, and if that is not specified OpenStackDataPlaneService.Name is
used.

This removes the coupling between the service name and the mount path so
that custom services can use the same ansible content, and the ansible
content can use a single consistent path.

Jira: https://issues.redhat.com/browse/OSPRH-7112
Related-To: https://github.com/openstack-k8s-operators/edpm-ansible/pull/658
Signed-off-by: James Slagle <jslagle@redhat.com>