Correct comment for OSDPS 'Play' object
This change fixes the terminology used to describe the 'Play' object of the OpenStackDataPlaneService.
This object is not an in-line 'playbook', but rather an individual Ansible 'Play'. The two are differentiated
by their type. A Ansible Playbook is a list of Plays, where a Play is a dictionary object as per:

https://github.com/ansible/ansible/blob/10f9b8e6554e024e3561170153b8e7fde5e7e4fb/test/units/playbook/test_play.py\#L48

Signed-off-by: Brendan Shephard <bshephar@redhat.com>