Add EDPMServiceType, Drop EDPMServiceName
Adds a top level EDPMServiceType field and drops EDPMServiceName.
EDPMServiceType is used to construct the volume mount paths for TLS
certs and CA certs, and should correspond to the ansible role name used
to managing the service. When not set, a defaulting webhook will default
the field to the the service name instead.

EDPMRoleServiceName is kept in case there is a need to override the path
for a specific TLS cert.

Signed-off-by: James Slagle <jslagle@redhat.com>