Add DataSources to OpenStackDataPlaneService
The DataSource struct is added to OpenStackDataPlaneService as an
additional way to specify ConfigMaps and Secrets.

This allows for specifying additional properties of the ConfigMaps and
Secrets, and for using the optional field on ConfigMapRef and SecretRef.
Deploying the service will no longer fail if the ConfigMap or Secret
doesn't exist when optional is true.

The existing ConfigMaps and Secrets fields are kept with the same
functionality for now for backwards compatibility with CI. Once CI
resources are updated to use OpenStackDataPlaneService.Spec.DataSources,
then these fields may be dropped.

AnsibleVarsFrom is also refactored to use the common DataSource code.

Signed-off-by: James Slagle <jslagle@redhat.com>