Rename Play parameter to PlaybookContents
We have a "Play" parameter in our Services that is leveraged to provide
a in-line Ansible playbook. This parameter is ultimately passed into the AEE container
using the RUNNER_PLAYBOOK env var. Given that an Ansible Play is a single component of
a playbook. This change rectifies our use of Ansible terminology and provides a more
accurate description of what the field expects to receive from the user.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>